피부 질환이 아닌 경우 응답 코드 추가 & float를 double형으로 수정

+++ src/main/java/com/example/capstone/aidisease/dto/DiagnosisResultResponse.java
@@ -9,7 +9,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class DiagnosisResultResponse {
     private String disease;
-    private float probability;
+    private double probability;
     private String treatment;
     private String source;
     private String imageUrl;
